+++ lib/src/sxwnl_dart_base.dart
@@ -1,4 +1,5 @@
 export 'astro_date_time.dart';
+export 'jie_qi.dart';
 export 'location.dart';
 export 'sxwnl/gan_zhi_calc.dart';
 export 'sxwnl/solar_lunar_pos.dart' show sALon, eLon, mLon;